+++ packages/react-ui/src/components/UserProfileInfo/UserProfileInfo.tsx
@@ -3,12 +3,13 @@ import {styled} from '@mui/material/styles';
 import {Box, Grid, Typography} from '@mui/material';
 import {defineMessages, useIntl} from 'react-intl';
 import {camelCase} from '@selfcommunity/utils';
-import {SCUserFields, SCUserType} from '@selfcommunity/types';
+import {SCUserType} from '@selfcommunity/types';
 import {useSCFetchUser} from '@selfcommunity/react-core';
 import {DEFAULT_FIELDS} from '../../constants/UserProfile';
 import UserProfileInfoSkeleton from './Skeleton';
 import classNames from 'classnames';
 import {useThemeProps} from '@mui/system';
+import {SCUserProfileFields} from '../../types';
 
 const messages = defineMessages({
   realName: {
@@ -80,7 +81,7 @@ export interface UserProfileInfoProps {
    * User fields to display in the profile
    * @default [real_name, date_joined, date_of_birth, website, description, bio]
    */
-  fields?: SCUserFields[];
+  fields?: SCUserProfileFields[];
 
   /**
    * Any other properties
@@ -128,8 +129,8 @@ export default function UserProfileInfo(inProps: UserProfileInfoProps): JSX.Elem
   // RENDER
   const renderField = (user, field) => {
     switch (field) {
-      case SCUserFields.DATE_JOINED:
-      case SCUserFields.DATE_OF_BIRTH:
+      case SCUserProfileFields.DATE_JOINED:
+      case SCUserProfileFields.DATE_OF_BIRTH:
         return `${intl.formatDate(user[field], {year: 'numeric', month: 'numeric', day: 'numeric'})}`;
       default:
         return user[field];