+++ packages/react-ui/src/types/index.ts
@@ -5,6 +5,7 @@ import {SCFeedObjectActivitiesType, SCFeedObjectTemplateType} from './feedObject
 import {SCNotificationObjectTemplateType} from './notification';
 import {SCBroadcastMessageTemplateType} from './broadcastMessage';
 import {SCHeaderMenuUrlsType} from './headerMenuUrls';
+import {SCUserProfileFields, SCUserProfileSettings} from './user';
 
 export {
   SCFeedWidgetType,
@@ -15,5 +16,7 @@ export {
   SCFeedObjectTemplateType,
   SCNotificationObjectTemplateType,
   SCBroadcastMessageTemplateType,
-  SCHeaderMenuUrlsType
+  SCHeaderMenuUrlsType,
+  SCUserProfileFields,
+  SCUserProfileSettings
 };