+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2-alpha.2](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.2-alpha.1...@selfcommunity/react-core@0.3.2-alpha.2) (2022-08-29)
+
+**Note:** Version bump only for package @selfcommunity/react-core
+
+
+
+
+
+## [0.3.2-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.2-alpha.0...@selfcommunity/react-core@0.3.2-alpha.1) (2022-08-25)
+
+**Note:** Version bump only for package @selfcommunity/react-core
+
+
+
+
+
+## [0.3.2-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.1...@selfcommunity/react-core@0.3.2-alpha.0) (2022-08-25)
+
+**Note:** Version bump only for package @selfcommunity/react-core
+
+
+
+
+
 ## [0.3.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.1-alpha.2...@selfcommunity/react-core@0.3.1) (2022-08-23)
 
 **Note:** Version bump only for package @selfcommunity/react-core