+++ packages/react-templates/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.37-alpha.2](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.37-alpha.1...@selfcommunity/react-templates@0.1.37-alpha.2) (2022-08-29)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.1.37-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.37-alpha.0...@selfcommunity/react-templates@0.1.37-alpha.1) (2022-08-25)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.1.37-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.36...@selfcommunity/react-templates@0.1.37-alpha.0) (2022-08-25)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
 ## [0.1.36](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.36-alpha.3...@selfcommunity/react-templates@0.1.36) (2022-08-23)
 
 **Note:** Version bump only for package @selfcommunity/react-templates